Use int64

+++ mini-vinci-be/go/common/result_models.go
@@ -11,5 +11,5 @@ type Result struct {
 	ProblemID       uint
 	LastSubmittedAt sql.NullTime
 	SubmissionCount uint
-	MaxScore        uint
+	MaxScore        int64
 }
